cleanup some code and move the details modal file
