+++ src/components/media/MediaCard.tsx
@@ -26,7 +26,8 @@ import { MediaBookmarkButton } from "./MediaBookmark";
 import { Button } from "../buttons/Button";
 import { IconPatch } from "../buttons/IconPatch";
 import { Icon, Icons } from "../Icon";
-import { DetailsModal, useModal } from "../overlays/Modal";
+import { DetailsModal } from "../overlays/DetailsModal";
+import { useModal } from "../overlays/Modal";
 
 export interface MediaCardProps {
   media: MediaItem;
