+++ src/components/player/internals/InfoButton.tsx
@@ -7,7 +7,8 @@ import {
   TMDBShowData,
 } from "@/backend/metadata/types/tmdb";
 import { Icons } from "@/components/Icon";
-import { DetailsModal, useModal } from "@/components/overlays/Modal";
+import { DetailsModal } from "@/components/overlays/DetailsModal";
+import { useModal } from "@/components/overlays/Modal";
 import { usePlayerStore } from "@/stores/player/store";
 
 import { VideoPlayerButton } from "./Button";
